Try to detect `execvp` failures for better error messages
$ PATH= ./show @
vcs: `git` error: No such file or directory

The current method being used (send SIGUSR1 to itself) can be
easily faked (eg. page output and then send SIGUSR1 to the `less`
child process), but it works good enough for now for most cases.

A more involved alternative could be to send SIGUSR1 to the parent
process, and then the parent process would catch that signal and
if the sender PID matched the child PID it just spawned, then that
would mean `execvp` failed.